+++ src/Document.php
@@ -120,58 +120,22 @@ final class Document
         return $this->encoding;
     }
 
-    public function content(): string
-    {
-        $children = $this->children->map(
-            static fn($child) => $child->toString(),
-        );
-
-        return Str::of('')->join($children)->toString();
-    }
-
-    public function toString(): string
-    {
-        $string = $this->tag();
-
-        $string .= $this->type->match(
-            static fn($type) => "\n".$type->toString(),
-            static fn() => '',
-        );
-
-        return $string."\n".$this->content();
-    }
-
-    public function asContent(): Content
+    public function asContent(Format $format = Format::pretty): Content
     {
         $writer = new \XMLWriter;
         /** @psalm-suppress ImpureMethodCall */
         $writer->openMemory();
-        /** @psalm-suppress ImpureMethodCall */
-        $writer->setIndent(true);
-        /** @psalm-suppress ImpureMethodCall */
-        $writer->setIndentString('    ');
+
+        if ($format === Format::pretty) {
+            /** @psalm-suppress ImpureMethodCall */
+            $writer->setIndent(true);
+            /** @psalm-suppress ImpureMethodCall */
+            $writer->setIndentString('    ');
+        }
 
         return Content::ofChunks(
             $this->render($writer),
         );
-    }
-
-    private function tag(): string
-    {
-        $writer = new \XMLWriter;
-        /** @psalm-suppress ImpureMethodCall */
-        $writer->openMemory();
-        /** @psalm-suppress ImpureMethodCall */
-        $writer->startDocument(
-            $this->version->toString(),
-            $this->encoding->match(
-                static fn($encoding) => $encoding->toString(),
-                static fn() => null,
-            ),
-        );
-
-        /** @psalm-suppress ImpureMethodCall */
-        return \trim($writer->outputMemory(), "\n");
     }
 
     /**