+++ tests/DocumentTest.php
@@ -11,6 +11,7 @@ use Innmind\Xml\{
     Element,
     Element\Name,
     Node,
+    Format,
 };
 use Innmind\Immutable\{
     Sequence,
@@ -82,32 +83,13 @@ class DocumentTest extends TestCase
         ));
     }
 
-    public function testContentWithoutChildren()
-    {
-        $this->assertSame(
-            '',
-            Document::of(Version::of(1), Maybe::nothing(), Maybe::nothing())->content(),
-        );
-    }
-
-    public function testContentWithChildren()
-    {
-        $this->assertSame(
-            '<foo></foo>',
-            Document::of(
-                Version::of(1),
-                Maybe::nothing(),
-                Maybe::nothing(),
-                Sequence::of(Element::of(Name::of('foo'))),
-            )->content(),
-        );
-    }
-
     public function testCast()
     {
         $this->assertSame(
             '<?xml version="2.1"?>'."\n",
-            Document::of(Version::of(2, 1), Maybe::nothing(), Maybe::nothing())->toString(),
+            Document::of(Version::of(2, 1), Maybe::nothing(), Maybe::nothing())
+                ->asContent(Format::inline)
+                ->toString(),
         );
         $this->assertSame(
             '<?xml version="2.1" encoding="UTF-8"?>'."\n",
@@ -115,7 +97,9 @@ class DocumentTest extends TestCase
                 Version::of(2, 1),
                 Maybe::nothing(),
                 Encoding::of('utf-8'),
-            )->toString(),
+            )
+                ->asContent(Format::inline)
+                ->toString(),
         );
         $this->assertSame(
             '<?xml version="2.1" encoding="UTF-8"?>'."\n".'<!DOCTYPE html>'."\n",
@@ -123,7 +107,9 @@ class DocumentTest extends TestCase
                 Version::of(2, 1),
                 Maybe::just(Type::of('html')),
                 Encoding::of('utf-8'),
-            )->toString(),
+            )
+                ->asContent(Format::inline)
+                ->toString(),
         );
         $this->assertSame(
             '<?xml version="2.1" encoding="UTF-8"?>'."\n".'<!DOCTYPE html>'."\n".'<foo/>',
@@ -132,7 +118,9 @@ class DocumentTest extends TestCase
                 Maybe::just(Type::of('html')),
                 Encoding::of('utf-8'),
                 Sequence::of(Element::selfClosing(Name::of('foo'))),
-            )->toString(),
+            )
+                ->asContent(Format::inline)
+                ->toString(),
         );
     }