+++ tests/Element/ElementTest.php
@@ -8,6 +8,7 @@ use Innmind\Xml\{
     Element\Name,
     Node,
     Attribute,
+    Format,
 };
 use Innmind\Immutable\{
     Map,
@@ -307,33 +308,13 @@ class ElementTest extends TestCase
         );
     }
 
-    public function testContentWithoutChildren()
-    {
-        $this->assertSame(
-            '',
-            Element::of(Name::of('foo'))->content(),
-        );
-    }
-
-    public function testContentWithChildren()
-    {
-        $node = Element::of(
-            Name::of('foo'),
-            null,
-            Sequence::of(Element::of(Name::of('bar'))),
-        );
-
-        $this->assertSame(
-            '<bar></bar>',
-            $node->content(),
-        );
-    }
-
     public function testCast()
     {
         $this->assertSame(
             '<foo></foo>',
-            Element::of(Name::of('foo'))->toString(),
+            Element::of(Name::of('foo'))
+                ->asContent(Format::inline)
+                ->toString(),
         );
         $this->assertSame(
             '<foo bar="baz" baz="foo"></foo>',
@@ -343,7 +324,9 @@ class ElementTest extends TestCase
                     Attribute::of('bar', 'baz'),
                     Attribute::of('baz', 'foo'),
                 ),
-            )->toString(),
+            )
+                ->asContent(Format::inline)
+                ->toString(),
         );
         $this->assertSame(
             '<foo bar="baz" baz="foo"><bar></bar><baz></baz></foo>',
@@ -357,7 +340,9 @@ class ElementTest extends TestCase
                     Element::of(Name::of('bar')),
                     Element::of(Name::of('baz')),
                 ),
-            )->toString(),
+            )
+                ->asContent(Format::inline)
+                ->toString(),
         );
     }