+++ tests/ReaderTest.php
@@ -9,6 +9,7 @@ use Innmind\Xml\{
     Element,
     Node,
     Attribute,
+    Format,
 };
 use Innmind\Filesystem\{
     Adapter\Filesystem,
@@ -63,7 +64,7 @@ XML;
             static fn() => null,
         );
 
-        $this->assertSame($xml, $node->toString());
+        $this->assertSame($xml, $node->asContent(Format::inline)->toString());
     }
 
     public function testReturnNothingWhenEmpty()
@@ -109,7 +110,7 @@ XML;
         );
         $this->assertSame(
             '<?xml-stylesheet type="text/xsl" href="/static/theatlantic/syndication/feeds/atom-to-html.6d0fbcbe7c3f.xsl" ?>',
-            $stylesheet->toString(),
+            $stylesheet->asContent()->toString(),
         );
     }